sup_qui_bug

+++ log/symfoLivre/src/Controller/UserController.php
@@ -50,21 +50,29 @@ class UserController extends AbstractController
         ]);
     }
     
-//    public function suppr(Request $request, User $user, UserRepository $userRepository): Response
-//        {
-//        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {        
-//             $userRepository->remove($user, true);
-//            
-//            //$manager= $this->getDoctrine()-getManager();
-//                
-//            //    $this->$manager->remove(User);
-//            //    $manager->flush();
-//        }
-//        return $this->redirectToRoute('user', [], Response::HTTP_SEE_OTHER);
-//    }
+    public function suppr(Request $request, User $user, UserRepository $userRepository): Response
+        {
+        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {        
+             $userRepository->remove($user, true);
+            
+            //$manager= $this->getDoctrine()-getManager();
+                
+            //    $this->$manager->remove(User);
+            //    $manager->flush();
+        }
+        return $this->redirectToRoute('user', [], Response::HTTP_SEE_OTHER);
+    }
 
-    public function suppr(Request $request, User $user, EntityManagerInterface $em): Response
+
+/**
+    public function suppr(Request $request, int $userId, EntityManagerInterface $em): Response
     {
+        $user = $this->getDoctrine()->getRepository(User::class)->find($userId);
+
+        if (!$user) {
+            throw $this->createNotFoundException('Utilisateur non trouvé');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
             $em->remove($user);
             $em->flush();
@@ -72,4 +80,5 @@ class UserController extends AbstractController
 
         return $this->redirectToRoute('user', [], Response::HTTP_SEE_OTHER);
     }
+*/    
 }
